move all pkevent stuff into datarecord
