+++ rdev/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mrgsolve
 Type: Package
-Version: 0.7.6.9019
+Version: 0.7.6.9020
 Title: Simulation from ODE-Based Population PK/PD and Systems Pharmacology
     Models
 Author: Kyle T. Baron [cre,aut,cph], Alan C. Hindmarsh [ctb], Linda R. Petzold
